+++ src/utils/ErrorMessages.js
@@ -1,5 +1,0 @@
-const ERROR_MESSAGES = {
-    FILE_DOESNT_EXIST: "The file \"{{FILE_URL}}\" doesn't exist."
-}
-
-export default ERROR_MESSAGES;